requirements.txt
# FastAPI and ASGI server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.10
asyncpg>=0.30.0
greenlet>=3.0.0

# Configuration and validation - Compatible with LangChain 0.3
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.5.0

# LLM providers and orchestration - LangChain 0.3 compatible
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-anthropic>=0.2.0,<0.3.0
langchain-mcp-adapters>=0.1.7
mcp-proxy>=0.8.0

# HTTP clients for MCP and external APIs
httpx>=0.25.0
aiohttp>=3.9.0

# AWS services
boto3>=1.34.0
botocore>=1.34.0

# Authentication and JWT
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Encryption
cryptography>=41.0.0

# Utilities
python-dotenv>=1.0.0
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-httpx>=0.26.0

# Development
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0 